requirements.txt
mcp[cli]>=1.0.0
requests>=2.31.0
urllib3>=2.0.0
librouteros>=3.2.0
python-dotenv>=1.0.0
pylint>=3.0.0
black>=24.0.0
